Fix 'tuplet' function

+++ smnp/library/function/tuplet.py
@@ -4,7 +4,7 @@ from smnp.type.model import Type
 
 _signature1 = varargSignature(ofTypes(Type.NOTE), ofTypes(Type.INTEGER), ofTypes(Type.INTEGER))
 def _function1(env, n, m, vararg):
-    t = [Type.note(arg.value.withDuration(arg.value.duration * n.value / m.value)) for arg in vararg]
+    t = [Type.note(arg.value.withDuration(int(arg.value.duration * n.value / m.value))) for arg in vararg]
     return Type.list(t).decompose()
